Fix build script and test code so it can complete 'mvn package' command and run on minikube with fractional executor's core

+++ server/src/main/scala/org/apache/livy/utils/SparkProcessBuilder.scala
@@ -104,7 +104,7 @@ class SparkProcessBuilder(livyConf: LivyConf) extends Logging {
     conf("spark.driver.cores", driverCores)
   }
 
-  def executorCores(executorCores: Int): SparkProcessBuilder = {
+  def executorCores(executorCores: Double): SparkProcessBuilder = {
     this.executorCores(executorCores.toString)
   }
